+++ changelog.md
@@ -1,4 +1,11 @@
 # Changelog
 
-**./test-parts/part-g-controller_files/controller-main.js**
-* Uncommented all tests.
+**./test-parts/part-h-api_requests/requests/req-admin_main.js**
+* Removed redundant string check from 'handleDefaultObject'
+	* testObjectPropertyContent
+
+---
+
+**./test-parts/part-h-api_requests/requests/req-alarm.js**
+* Removed redundant string checks from 'handleAvailable'
+	* testPropertyContents